# Optimize all functions for 2D arrays